Move RHEmailEventPersons to persons module
