+++ indico/modules/events/management/blueprint.py
@@ -18,7 +18,7 @@ from __future__ import unicode_literals
 
 from indico.modules.events import event_management_object_url_prefixes
 from indico.modules.events.management.controllers import (RHDeleteEvent, RHLockEvent, RHUnlockEvent,
-                                                          RHEmailEventPersons, RHShowNonInheriting)
+                                                          RHShowNonInheriting)
 from indico.web.flask.wrappers import IndicoBlueprint
 
 
@@ -29,7 +29,6 @@ _bp = IndicoBlueprint('event_management', __name__, template_folder='templates',
 _bp.add_url_rule('/delete', 'delete', RHDeleteEvent, methods=('GET', 'POST'))
 _bp.add_url_rule('/lock', 'lock', RHLockEvent, methods=('GET', 'POST'))
 _bp.add_url_rule('/unlock', 'unlock', RHUnlockEvent, methods=('POST',))
-_bp.add_url_rule('/manage/person-list/email', 'email_event_persons', RHEmailEventPersons, methods=('POST',))
 
 for object_type, prefixes in event_management_object_url_prefixes.iteritems():
     if object_type == 'subcontribution':
