+++ indico/modules/events/persons/blueprint.py
@@ -16,7 +16,7 @@
 
 from __future__ import unicode_literals
 
-from indico.modules.events.persons.controllers import RHPersonsList, RHPendingPersonsList
+from indico.modules.events.persons.controllers import RHPersonsList, RHPendingPersonsList, RHEmailEventPersons
 from indico.web.flask.wrappers import IndicoBlueprint
 
 _bp = IndicoBlueprint('persons', __name__, template_folder='templates', virtual_template_folder='events/persons',
@@ -24,3 +24,5 @@ _bp = IndicoBlueprint('persons', __name__, template_folder='templates', virtual_
 
 _bp.add_url_rule('/persons/', 'person_list', RHPersonsList)
 _bp.add_url_rule('/persons/pending', 'pending_persons_list', RHPendingPersonsList)
+_bp.add_url_rule('/persons/email', 'email_event_persons', RHEmailEventPersons, methods=('POST',))
+
